requirements-dev.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
google-api-core==2.8.0
google-api-python-client==2.125.0
google-api-python-client-stubs==1.25.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.56.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
oauthlib==3.2.2
pip==24.0
protobuf==3.12.4
py-cord==2.4.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==65.5.0
six==1.16.0
SQLAlchemy==2.0.27
types-httplib2==0.22.0.20240310
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240203
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
